refactor: TextEditor.tsx 코드 추상화하여 분리
