+++ src/app/testeditor/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import TextEditor from './TextEditor';
+import TextEditor from '../../components/common/TextEditor/TextEditor';
 
 const TestEditor = () => {
   return (
